requirements.txt
Django
Pillow
MySql-python
django-endless-pagination